requirements.txt
aiosqlite==0.16.1
anyio==3.2.1
asgiref==3.4.1
asyncpg==0.23.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.66.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
packaging==21.0
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
